fixed the issues
