+++ entgql/interfaces/repo/token_revocation.go
@@ -1,0 +1,10 @@
+package repo
+
+import (
+	"context"
+	"time"
+)
+
+type TokenRevocation interface {
+	RevokeSubjectTokens(ctx context.Context, subjectEmail, subjectType string, revokedAt time.Time) error
+}
